requirements.txt
fastapi
uvicorn
joblib
pandas
pydantic
groq
python-dotenv